CompanyController, GameController make REST

+++ src/main/java/com/web/portal/repository/CompanyRepository.java
@@ -4,5 +4,5 @@ import com.web.portal.models.Company;
 import org.springframework.data.repository.CrudRepository;
 
 public interface CompanyRepository extends CrudRepository<Company, Long> {
-    Company findFirstById(int id);
+    Company findFirstById(long id);
 }
